[#181294869] ruby 2.7.5, add nokogiri deps, make in parallel, compile geckodriver
This increases the final image size, from 500mb to 640MB.
However this lowers the local build-time with 4 minutes (or 35seconds if using alpine edge prebuilt).
This can also be reused for javascript tests in control center and gud (spectre for example), currently non-existant.

The ruby upgrade is required to get a newer version of alpine,
v3.15 is required for a fresh enough version of rust to build geckodriver.